MillionHeroAssistant/requirements.txt
--index-url https://pypi.tuna.tsinghua.edu.cn/simple
beautifulsoup4==4.6.0
certifi==2017.11.5
chardet==3.0.4
future==0.16.0
idna==2.6
jieba==0.39
lxml==4.1.1
macholib==1.9
pefile==2017.11.5
Pillow==5.0.0
PyYAML==3.12
requests==2.18.4
scikit-image==0.13.1
scipy==1.0.0
selenium==3.8.1
six==1.11.0
urllib3==1.22
numpy==1.14.0
baidu-aip==2.0.0.1
websocket-client==0.46.0
